pyproject.toml
[project]
name = "ppt-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.3",
    "openai>=1.52.2",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "python-pptx>=1.0.2",
]
